Add files via upload
This Project Collects Reviews from Users who aren't Satisfied by the Product. These Reviews are then taken to the Notice of the Development Team, to make changes and improvements to the Existing Product . All the Negative Reviews Help in Making the Product compete With other competitives.
To Separate the Negative Reviews from Positive Reviews , we use NLP with Python.